Update random extension
- Update changelog, reference #9609
- Feature 9609 add option to set Default action (copy/paste)
- Refactoring code to make it more maintainable
- Revert faker update, lint code
- Update dependencies & lint/prettier config
- Initial commit
