+++ extensions/random-data-generator/.prettierrc
@@ -1,4 +1,20 @@
 {
   "printWidth": 120,
-  "singleQuote": true
+  "singleQuote": false,
+  "importOrder": [
+    "<THIRD_PARTY_MODULES>",
+    "^@raycast/(.*)$",
+    "^@/api/$",
+    "^@/context/$",
+    "^@/hooks/$",
+    "^@/utils/$",
+    "^@/pages/$",
+    "^@/(.*)$",
+    "^[./]"
+  ],
+  "importOrderSeparation": true,
+  "importOrderSortSpecifiers": true,
+  "plugins": [
+    "@trivago/prettier-plugin-sort-imports"
+  ]
 }
