+++ extensions/random-data-generator/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Random Data Generator Changelog
 
-## [Quicklinks & CUID]
+## [Refactoring & Preferences] - 2024-02-04
+
+- Refactored code to make it more maintainable
+- Added the ability to set preference: `Default Action` to `Copy` or `Save` (issue #9609)
+- Bump dependencies (left @fakerjs/faker at the old version, locales have breaking changes in v8)
+
+## [Quicklinks & CUID] - 2023-10-23
 
 - Added CUIDs as a possible data type to generate
 - Add the ability to create quicklinks for copy-pasting specific random data
