+++ extensions/random-data-generator/package.json
@@ -8,7 +8,8 @@
   "contributors": [
     "Calvein",
     "pernielsentikaer",
-    "vimtor"
+    "vimtor",
+    "j3lte"
   ],
   "license": "MIT",
   "categories": [
@@ -32,30 +33,52 @@
       "mode": "no-view"
     }
   ],
+  "preferences": [
+    {
+      "name": "defaultAction",
+      "type": "dropdown",
+      "data": [
+        {
+          "title": "Copy to Clipboard",
+          "value": "copy"
+        },
+        {
+          "title": "Paste in Active App",
+          "value": "paste"
+        }
+      ],
+      "title": "Default Action",
+      "description": "Choose the default action for the generated data.",
+      "required": false,
+      "default": "copy",
+      "value": "copy"
+    }
+  ],
   "dependencies": {
     "@bugsnag/cuid": "^3.0.2",
     "@faker-js/faker": "^7.6.0",
-    "@raycast/api": "^1.60.1",
+    "@raycast/api": "^1.66.2",
     "is-url": "^1.2.4",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
-    "@types/bugsnag__cuid": "^3.0.1",
-    "@types/is-url": "^1.2.30",
-    "@types/lodash": "^4.14.186",
-    "@types/node": "^18.11.0",
-    "@types/react": "^18.0.21",
-    "@typescript-eslint/eslint-plugin": "^5.40.0",
-    "@typescript-eslint/parser": "^5.40.0",
-    "eslint": "^8.25.0",
-    "eslint-config-prettier": "^8.5.0",
-    "react-devtools": "^4.26.1",
-    "typescript": "^4.8.4"
+    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
+    "@types/bugsnag__cuid": "^3.0.2",
+    "@types/is-url": "^1.2.32",
+    "@types/lodash": "^4.14.202",
+    "@types/node": "^20.11.16",
+    "@types/react": "^18.2.52",
+    "@raycast/eslint-config": "^1.0.8",
+    "eslint": "^8.56.0",
+    "react-devtools": "^4.28.4",
+    "prettier": "^3.2.5",
+    "typescript": "^5.3.3"
   },
   "scripts": {
     "build": "ray build -e dist",
     "dev": "ray develop",
     "fix-lint": "ray lint --fix",
-    "lint": "ray lint"
+    "lint": "ray lint",
+    "publish": "npx @raycast/api@latest publish"
   }
 }
