+++ extensions/random-data-generator/src/hooks/usePreferences.ts
@@ -1,0 +1,13 @@
+import { getPreferenceValues } from "@raycast/api";
+
+interface Preferences {
+  defaultAction: "copy" | "paste";
+}
+
+function usePreferences(): Preferences;
+function usePreferences<T extends keyof Preferences>(key: T): Preferences[T];
+function usePreferences<T extends keyof Preferences>(key?: T) {
+  return key ? getPreferenceValues()[key] : getPreferenceValues();
+}
+
+export default usePreferences;
